SKIP_LOCKED in GP (#185)
Adds skip locked feature to GPDB. 
ANALYZE SKIPLOCKED will process all leaf partitions of tables, which are NOT currently locked. 
Note that this feature does not protect from concurrent lock acquisition, so semantics is "best-effort".
Note that non-leaf partitions will not be updated even if no lock is actually held.