Add fts maintenance mode for cluster/DC maintenance (#251)
This change adds a gp_fts_maintenance GUC and allows turning FTS off completely. Useful when expecting problems in DC or recovering from a DC failure.

When FTS is turned off, the failures are handled this way:

a mirror failure will result in read-only cluster, but FTS won't change gp_segment_configuration. If a mirror[s] is started either with pg_ctl or with cluster restart, no gprecoverseg is needed - the cluster will just continue working as if nothing happened
a primary segment failure results in cluster that is not available for both reading and writing. Again, no changed in gp_segment_configuration. Again, either restart or pg_ctl can start the failed segments and cluster will continue operating
if there is a double fault during fts maintenance (i.e., mirror and primary with the same content are down), cluster can be fixed with either manual start of segments, or whole cluster restart. HOWEVER if double fault is still present during cluster start, gpstart will fail. Which is the expected begaviour. Again, no changes in gp_segment_configuration will occur.
The idea is that as soon as the infrastructure is stable, we can turn maintenance off and run gprecoverseg -Fa if necessary.