Make pgaudit tests pass (#270)
Now pgaudit works on master only. So we don't receive messages from segments,
and the messages order is stable.
Fix use after free for auditEventStack->auditEvent.objectName.
Add DISTRIBUTED to the CREATE TABLE queries and RESOURCE QUEUE to CREATE ROLE to
get rid of Greenplum specific warnings.
Make changes in pg_hba.conf to login as users which are created in the tests.
Wrap the current user name in quotation marks, because it can contain symbols
which broke the tests, for example '-'.
Remove queries which Greenplum cannot execute: queries with more then one
writing slice and triggers updating each row.