Fix subplans with correlated functions or master-only or replicated tables (#89)
The add_rte_to_flat_rtable function places a RangeTblEntry into the
glob->finalrtable table with a zeroed list of functions. The
ParallelizeCorrelatedSubPlanMutator function uses ctx->rtable, which is
populated by the root->glob->finalrtable function in the
ParallelizeCorrelatedSubPlan function. Therefore, in the
ParallelizeCorrelatedSubPlanMutator function, the rte->functions list is always
empty. Therefore, the old condition for checking whether functions are
correlated did not work. Now, in the ParallelizeCorrelatedSubPlanMutator
function, functions are available in the fscan->functions list, which is where
they are taken for correlation and location checking. Correlated non-any
functions are prohibited in subplans. For other combinations, we now add
broadcasting and materialization.

Another issue arose with correlated subplans with master-only or replicated
tables. The existing condition checked the scan locus only, without checking
the requested locus. As a result, the table scan was performed on segments,
which is unacceptable for the gp_segment_configuration master-only table. Now,
if the requested locus for such tables and the scan locus are not equal to the
entry, we again add broadcasting and materialization.

Co-authored-by: Georgy Shelkovy <g.shelkovy@arenadata.io>
Co-authored-by: Maxim Michkov <m.michkov@arenadata.io>

Ticket: ADBDEV-6884